+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Fixed
 - line item totals in order confirmation mails
 - tax calculation for international orders
+- add order error with missing currency factor
 
 ## [2.9.78] - 2018-04-10
 ### Fixed